Создал директории для каждой из задач в домашке.
Разбил классы по листингам.
